add ObjectGN for all class using geometry nodes
